feat: setup to read problem input file
The goal is to have a single `index.html` to which we can input any of
the Advent of Code files and have it return the proper response. As I am
using this to learn javascript, this will probably not be the cleanest
code. Due to CORS policy I can't use modules since I want it to all run
client-side. I'll need to replicate the file reading logic for each
`.js` solution.